refactor covert.dart
